+++ .idea/gradle.xml
@@ -32,9 +32,9 @@
             <option value="$PROJECT_DIR$/core/repository" />
             <option value="$PROJECT_DIR$/core/ui" />
             <option value="$PROJECT_DIR$/core/util" />
-            <option value="$PROJECT_DIR$/feature/featurea" />
             <option value="$PROJECT_DIR$/feature/featureb" />
             <option value="$PROJECT_DIR$/feature/featurec" />
+            <option value="$PROJECT_DIR$/feature/rss" />
           </set>
         </option>
       </GradleProjectSettings>
