+++ app/src/main/java/com/github/goutarouh/androidsampleapp/MainBottomNavigation.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.github.goutarouh.androidsampleapp.feature.featurea.RSS_HOME_ROUTE
+import com.github.goutarouh.androidsampleapp.feature.rss.rsshome.RSS_HOME_ROUTE
 import com.github.goutarouh.androidsampleapp.feature.featureb.featureBRoute
 import com.github.goutarouh.androidsampleapp.feature.featurec.featureCRoute
