--- feature/featurea/build.gradle
@@ -1,43 +1,0 @@
-plugins {
-    id "androidsampleapp.android.library"
-    id "androidsampleapp.android.hilt"
-}
-
-android {
-    namespace 'com.github.goutarouh.androidsampleapp.feature.featurea'
-
-    defaultConfig {
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-        consumerProguardFiles "consumer-rules.pro"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
-        }
-    }
-    buildFeatures {
-        compose true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion '1.1.1'
-    }
-}
-
-dependencies {
-
-    implementation 'androidx.core:core-ktx:1.7.0'
-    implementation 'androidx.appcompat:appcompat:1.5.1'
-    implementation 'com.google.android.material:material:1.7.0'
-    testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.4'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.0'
-
-    implementation 'androidx.hilt:hilt-navigation-compose:1.0.0'
-
-    implementation "androidx.navigation:navigation-compose:$compose_navigation_version"
-
-    implementation(project(":ui"))
-    implementation(project(":repository"))
-}
