--- feature/featurea/src/main/java/com/github/goutarouh/androidsampleapp/feature/featurea/rssitemlist/RssItemListScreenUiState.kt
@@ -1,15 +1,0 @@
-package com.github.goutarouh.androidsampleapp.feature.featurea.rssitemlist
-
-import com.github.goutarouh.androidsampleapp.core.repository.model.rss.Rss
-
-sealed interface RssItemListScreenUiState {
-
-    object Loading: RssItemListScreenUiState
-    data class Error(
-        val e: Exception
-    ): RssItemListScreenUiState
-    data class Success(
-        val rss: Rss
-    ): RssItemListScreenUiState
-
-}
