+++ feature/rss/src/main/AndroidManifest.xml
