+++ feature/rss/src/main/java/com/github/goutarouh/androidsampleapp/feature/rss/rsshome/RssHomeScreenUiState.kt
@@ -1,0 +1,9 @@
+package com.github.goutarouh.androidsampleapp.feature.rss.rsshome
+
+import com.github.goutarouh.androidsampleapp.core.repository.model.rss.Rss
+
+sealed interface RssHomeScreenUiState {
+    object Initial: RssHomeScreenUiState
+    data class Error(val e: Exception): RssHomeScreenUiState
+    data class Success(val rssList: List<Rss>): RssHomeScreenUiState
+}
